added stack problem
