+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+## feel free to add on any program under suitable folder following the guidelines given below
+the program must be optimized\
+it should have comments wherever needed\
+the name of the program should give an idea what the program is about
+
+## happy coding :)
+
+
+
