+++ Python /linked list/FindAnElementInDoublyLinkedList.py
